Added CI module
Added the CI module. This is paired with an existing Jenkins
installation with a job that executes the tests. Not everyone can do the
excercise simultaneously, though. The excercise must be controlled
during the lesson. In the lesson, the Jenkins example is also shown to
students.

Harmonized DevOps part a bit with this module.